integration: extract pager.rs tests
Tests:
- test_iter_works_when_retry_policy_returns_ignore_write_error()
- test_iter_methods_with_modification_statements()
are extracted into `session/pager.rs`.
