+++ scylla/tests/integration/session/mod.rs
@@ -2,6 +2,7 @@ mod authenticate;
 mod caching_session;
 mod history;
 mod new_session;
+mod pager;
 mod retries;
 mod self_identity;
 #[expect(clippy::module_inception)]
