+++ scylla/tests/integration/session/session.rs
@@ -7,21 +7,14 @@ use scylla::client::execution_profile::ExecutionProfile;
 use scylla::client::session::Session;
 use scylla::errors::OperationType;
 use scylla::errors::{DbError, ExecutionError, RequestAttemptError};
-use scylla::policies::retry::{RequestInfo, RetryDecision, RetryPolicy, RetrySession};
 use scylla::routing::partitioner::PartitionerName;
 use scylla::routing::Token;
 use scylla::serialize::row::SerializeRow;
 use scylla::statement::prepared::PreparedStatement;
 use scylla::statement::unprepared::Statement;
-use scylla::statement::Consistency;
-use scylla::value::Row;
 
 use assert_matches::assert_matches;
-use futures::{StreamExt as _, TryStreamExt};
 use std::collections::{BTreeMap, HashMap};
-use std::fmt::Debug;
-use std::sync::atomic::{AtomicBool, Ordering};
-use std::sync::Arc;
 use std::time::Duration;
 use std::vec;
 
@@ -473,133 +466,4 @@ async fn test_rate_limit_exceeded_exception() {
         }
         err => panic!("Unexpected error type received: {:?}", err),
     }
-}
-
-// Reproduces the problem with execute_iter mentioned in #608.
-#[tokio::test]
-async fn test_iter_works_when_retry_policy_returns_ignore_write_error() {
-    setup_tracing();
-    // It's difficult to reproduce the issue with a real downgrading consistency policy,
-    // as it would require triggering a WriteTimeout. We just need the policy
-    // to return IgnoreWriteError, so we will trigger a different error
-    // and use a custom retry policy which returns IgnoreWriteError.
-    let retried_flag = Arc::new(AtomicBool::new(false));
-
-    #[derive(Debug)]
-    struct MyRetryPolicy(Arc<AtomicBool>);
-    impl RetryPolicy for MyRetryPolicy {
-        fn new_session(&self) -> Box<dyn RetrySession> {
-            Box::new(MyRetrySession(self.0.clone()))
-        }
-    }
-
-    struct MyRetrySession(Arc<AtomicBool>);
-    impl RetrySession for MyRetrySession {
-        fn decide_should_retry(&mut self, _: RequestInfo) -> RetryDecision {
-            self.0.store(true, Ordering::Relaxed);
-            RetryDecision::IgnoreWriteError
-        }
-        fn reset(&mut self) {}
-    }
-
-    let handle = ExecutionProfile::builder()
-        .consistency(Consistency::All)
-        .retry_policy(Arc::new(MyRetryPolicy(retried_flag.clone())))
-        .build()
-        .into_handle();
-
-    let session = create_new_session_builder()
-        .default_execution_profile_handle(handle)
-        .build()
-        .await
-        .unwrap();
-
-    // Create a keyspace with replication factor that is larger than the cluster size
-    let cluster_size = session.get_cluster_state().get_nodes_info().len();
-    let ks = unique_keyspace_name();
-    let mut create_ks = format!("CREATE KEYSPACE {} WITH REPLICATION = {{'class': 'NetworkTopologyStrategy', 'replication_factor': {}}}", ks, cluster_size + 1);
-    if scylla_supports_tablets(&session).await {
-        create_ks += " and TABLETS = { 'enabled': false}";
-    }
-    session.ddl(create_ks).await.unwrap();
-    session.use_keyspace(ks, true).await.unwrap();
-    session
-        .ddl("CREATE TABLE t (pk int PRIMARY KEY, v int)")
-        .await
-        .unwrap();
-
-    assert!(!retried_flag.load(Ordering::Relaxed));
-    // Try to write something to the new table - it should fail and the policy
-    // will tell us to ignore the error
-    let mut stream = session
-        .query_iter("INSERT INTO t (pk v) VALUES (1, 2)", ())
-        .await
-        .unwrap()
-        .rows_stream::<Row>()
-        .unwrap();
-
-    assert!(retried_flag.load(Ordering::Relaxed));
-    while stream.try_next().await.unwrap().is_some() {}
-
-    retried_flag.store(false, Ordering::Relaxed);
-    // Try the same with execute_iter()
-    let p = session
-        .prepare("INSERT INTO t (pk, v) VALUES (?, ?)")
-        .await
-        .unwrap();
-    let mut iter = session
-        .execute_iter(p, (1, 2))
-        .await
-        .unwrap()
-        .rows_stream::<Row>()
-        .unwrap()
-        .into_stream();
-
-    assert!(retried_flag.load(Ordering::Relaxed));
-    while iter.try_next().await.unwrap().is_some() {}
-}
-
-#[tokio::test]
-async fn test_iter_methods_with_modification_statements() {
-    let session = create_new_session_builder().build().await.unwrap();
-    let ks = unique_keyspace_name();
-
-    session.ddl(format!("CREATE KEYSPACE IF NOT EXISTS {} WITH REPLICATION = {{'class' : 'NetworkTopologyStrategy', 'replication_factor' : 1}}", ks)).await.unwrap();
-    session
-        .ddl(format!(
-            "CREATE TABLE IF NOT EXISTS {}.t (a int, b int, c text, primary key (a, b))",
-            ks
-        ))
-        .await
-        .unwrap();
-
-    let mut query = Statement::from(format!(
-        "INSERT INTO {}.t (a, b, c) VALUES (1, 2, 'abc')",
-        ks
-    ));
-    query.set_tracing(true);
-    let mut rows_stream = session
-        .query_iter(query, &[])
-        .await
-        .unwrap()
-        .rows_stream::<Row>()
-        .unwrap();
-    rows_stream.next().await.ok_or(()).unwrap_err(); // assert empty
-    assert!(!rows_stream.tracing_ids().is_empty());
-
-    let prepared_statement = session
-        .prepare(format!("INSERT INTO {}.t (a, b, c) VALUES (?, ?, ?)", ks))
-        .await
-        .unwrap();
-    let query_pager = session
-        .execute_iter(prepared_statement, (2, 3, "cba"))
-        .await
-        .unwrap();
-    query_pager
-        .rows_stream::<()>()
-        .unwrap()
-        .next()
-        .await
-        .ok_or(())
-        .unwrap_err(); // assert empty
 }
